Página de sign-up en modalidad SPA. Funcionalidad de enmascarar y desenmascarar contraseña, dos campos para ingresar contraseña, se desactiva el botón de crear cuenta hasta que el usuario ingresa la misma contraseña en ambos campos.
